requirements.txt
# Basic Flask dependencies
flask==2.3.3
requests>=2.28.2
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Google-related libraries
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib==0.4.6
google-auth>=2.22.0
google-generativeai>=0.3.1

# Audio processing
whisper>=1.1.10
pyannote.audio>=2.1.1
PyPDF2>=3.0.1

# NEW: Audio preprocessing dependencies
numpy>=1.23.5
librosa>=0.10.1
soundfile>=0.12.1

openai-whisper
# Add any other dependencies here
